requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.8
outcome==1.3.0.post0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-decouple==3.8
redis==5.0.8
requests==2.32.3
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
